requirements.txt
# Streamlit app framework
streamlit

# PDF parsing
pypdf

# Text chunking
langchain

# Sentence Embedding models
sentence-transformers

# FAISS for vector search
faiss-cpu

# OpenAI LLM API
openai

# Environment variables (.env file loader)
python-dotenv

# Pickle is built-in, no need to install
# Numpy needed for FAISS
numpy
